requirements.txt
streamlit
google-api-python-client
python-dotenv
psycopg2-binary
pytest
